+++ conda/cli/conda_argparse.py
@@ -126,6 +126,11 @@ def add_parser_create_install_update(p):
         help="Do not install dependencies.",
     )
     p.add_argument(
+        "--only-deps",
+        action="store_true",
+        help="Only install dependencies.",
+    )
+    p.add_argument(
         '-m', "--mkdir",
         action="store_true",
         help="Create the environment directory if necessary.",
@@ -138,6 +143,7 @@ def add_parser_create_install_update(p):
     add_parser_prefix(p)
     add_parser_quiet(p)
     add_parser_copy(p)
+    add_parser_insecure(p)
     p.add_argument(
         "--alt-hint",
         action="store_true",
@@ -389,4 +395,14 @@ def add_parser_no_use_index_cache(p):
         default=True,
         dest="use_index_cache",
         help="Force fetching of channel index files.",
+    )
+
+
+def add_parser_insecure(p):
+    p.add_argument(
+        "-k", "--insecure",
+        action="store_false",
+        default=NULL,
+        help="Allow conda to perform \"insecure\" SSL connections and transfers."
+             "Equivalent to setting 'ssl_verify' to 'false'."
     )