+++ conda/cli/install.py
@@ -10,7 +10,6 @@ from difflib import get_close_matches
 import logging
 import os
 from os.path import abspath, basename, exists, isdir, join
-import re
 
 from . import common
 from .._vendor.auxlib.ish import dals
@@ -36,8 +35,6 @@ log = logging.getLogger(__name__)
 def check_prefix(prefix, json=False):
     name = basename(prefix)
     error = None
-    if name.startswith('.'):
-        error = "environment name cannot start with '.': %s" % name
     if name == ROOT_ENV_NAME:
         error = "'%s' is a reserved environment name" % name
     if exists(prefix):
@@ -316,19 +313,15 @@ def install(args, parser, command='install'):
         if e.args and 'could not import' in e.args[0]:
             raise CondaImportError(text_type(e))
         raise
+
+    if unlink_link_transaction.nothing_to_do and not newenv:
+        if context.json:
+            common.stdout_json_success(message='All requested packages already installed.')
+        else:
+            print('\n# All requested packages already installed.\n')
+        return
+
     if not context.json:
-        if unlink_link_transaction.nothing_to_do and not newenv:
-            from .main_list import print_packages
-
-            if not context.json:
-                spec_regex = r'^(%s)$' % re.escape('|'.join(s.split()[0] for s in ospecs))
-                print('\n# All requested packages already installed.')
-                print_packages(prefix, spec_regex)
-            else:
-                common.stdout_json_success(
-                    message='All requested packages already installed.')
-            return
-
         unlink_link_transaction.display_actions(progressive_fetch_extract)
         common.confirm_yn(args)