+++ conda/core/repodata.py
@@ -17,7 +17,7 @@ from time import time
 import warnings
 
 from requests import ConnectionError, HTTPError
-from requests.exceptions import SSLError
+from requests.exceptions import InvalidSchema, SSLError
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
 
 from .. import CondaError, iteritems
@@ -26,10 +26,10 @@ from .._vendor.auxlib.ish import dals
 from .._vendor.auxlib.logz import stringify
 from ..base.constants import CONDA_HOMEPAGE_URL
 from ..base.context import context
-from ..common.compat import ensure_binary, ensure_text_type, ensure_unicode
+from ..common.compat import ensure_binary, ensure_text_type, ensure_unicode, text_type
 from ..common.url import join_url, maybe_unquote
 from ..core.package_cache import PackageCache
-from ..exceptions import CondaHTTPError, CondaIndexError
+from ..exceptions import CondaDependencyError, CondaHTTPError, CondaIndexError
 from ..gateways.connection import CondaSession
 from ..gateways.disk.delete import rm_rf
 from ..gateways.disk.update import touch
@@ -154,9 +154,17 @@ def fetch_repodata_remote_request(session, url, etag, mod_stamp):
         add_http_value_to_dict(resp, 'Last-Modified', fetched_repodata, '_mod')
         add_http_value_to_dict(resp, 'Cache-Control', fetched_repodata, '_cache_control')
         return fetched_repodata
-
-    except ValueError as e:
-        raise CondaIndexError("Invalid index file: {0}: {1}".format(join_url(url, filename), e))
+    except InvalidSchema as e:
+        if 'SOCKS' in text_type(e):
+            message = dals("""
+            Requests has identified that your current working environment is configured
+            to use a SOCKS proxy, but pysocks is not installed.  To proceed, remove your
+            proxy configuration, run `conda install pysocks`, and then you can re-enable
+            your proxy configuration.
+            """)
+            raise CondaDependencyError(message)
+        else:
+            raise
 
     except (ConnectionError, HTTPError, SSLError) as e:
         # status_code might not exist on SSLError
@@ -305,6 +313,9 @@ def fetch_repodata_remote_request(session, url, etag, mod_stamp):
                              getattr(e.response, 'elapsed', None),
                              e.response)
 
+    except ValueError as e:
+        raise CondaIndexError("Invalid index file: {0}: {1}".format(join_url(url, filename), e))
+
 
 def write_pickled_repodata(cache_path, repodata):
     # Don't bother to pickle empty channels
@@ -413,7 +424,7 @@ def fetch_repodata(url, schannel, priority,
         mtime = getmtime(cache_path)
     except (IOError, OSError):
         log.debug("No local cache found for %s at %s", url, cache_path)
-        if use_cache:
+        if use_cache or (context.offline and not url.startswith('file://')):
             return {'packages': {}}
         else:
             mod_etag_headers = {}