+++ conda/gateways/disk/create.py
@@ -8,7 +8,7 @@ from logging import getLogger
 import os
 from os import X_OK, access, makedirs
 from os.path import basename, dirname, isdir, isfile, join, splitext
-import shutil
+from shutil import copy as shutil_copy, copystat
 import sys
 import tarfile
 import traceback
@@ -200,7 +200,7 @@ def create_hard_link_or_copy(src, dst):
         link(src, dst)
     except (IOError, OSError):
         log.info('hard link failed, so copying %s => %s', src, dst)
-        shutil.copy2(src, dst)
+        _do_copy(src, dst)
 
 
 def _is_unix_executable_using_ORIGIN(path):
@@ -218,6 +218,7 @@ def _do_softlink(src, dst):
         # A future optimization will be to copy code from @mingwandroid's virtualenv patch.
         copy(src, dst)
     else:
+        log.trace("soft linking %s => %s", src, dst)
         symlink(src, dst)
 
 
@@ -238,9 +239,21 @@ def copy(src, dst):
         src_points_to = readlink(src)
         if not src_points_to.startswith('/'):
             # copy relative symlinks as symlinks
+            log.trace("soft linking %s => %s", src, dst)
             symlink(src_points_to, dst)
             return
-    shutil.copy2(src, dst)
+    _do_copy(src, dst)
+
+
+def _do_copy(src, dst):
+    log.trace("copying %s => %s", src, dst)
+    shutil_copy(src, dst)
+    try:
+        copystat(src, dst)
+    except (IOError, OSError) as e:  # pragma: no cover
+        # shutil.copystat gives a permission denied when using the os.setxattr function
+        # on the security.selinux property.
+        log.debug('%r', e)
 
 
 def create_link(src, dst, link_type=LinkType.hardlink, force=False):
@@ -268,6 +281,7 @@ def create_link(src, dst, link_type=LinkType.hardlink, force=False):
         if isdir(src):
             raise CondaError("Cannot hard link a directory. %s" % src)
         try:
+            log.trace("hard linking %s => %s", src, dst)
             link(src, dst)
         except (IOError, OSError) as e:
             log.debug("%r", e)